+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **Indonesian (id) language support** ([#137](https://github.com/speedyk-005/yasbd-lib/pull/137)).
 - **Slovak (sk) language support** ([#129](https://github.com/speedyk-005/yasbd-lib/pull/129)).
 - **Vietnamese (vi) language support** ([#131](https://github.com/speedyk-005/yasbd-lib/pull/131)).
 - **Persian (fa) language support** ([#128](https://github.com/speedyk-005/yasbd-lib/pull/128)).